refactor: consolidate codemirror md extensions in codemirror-markdown package
