+++ sites/gaios/src/tools/index.ts
@@ -37,8 +37,6 @@ import { plugins as codemirrorEmbedPlugins } from "@grjte/codemirror-embed";
 // @ts-expect-error no types
 import { plugins as codemirrorMarkdownPlugins } from "@grjte/codemirror-markdown";
 // @ts-expect-error no types
-import { plugins as markdownLinksPlugins } from "@grjte/codemirror-md-links";
-// @ts-expect-error no types
 import { plugins as syncIndicatorPlugins } from "@tiny-patchwork/sync-indicator";
 // @ts-expect-error no types
 import { plugins as commandsPlugins } from "@orion/commands";
@@ -64,7 +62,6 @@ export const plugins: Plugin<any>[] = [
   ...syncIndicatorPlugins,
   ...codemirrorBasePlugins,
   ...codemirrorMarkdownPlugins,
-  ...markdownLinksPlugins,
   ...codemirrorEmbedPlugins,
   ...tenfoldPlugins,
   {
