+++ tools/codemirror/codemirror-markdown/src/main.tsx
@@ -2,16 +2,6 @@ import type { Extension } from "@codemirror/state";
 
 export const plugins = [
   {
-    type: "codemirror:extension",
-    id: "codemirror-markdown",
-    name: "Markdown",
-    supportedDatatypes: ["essay", "markdown"],
-    async load(): Promise<Extension> {
-      const { markdownExtensions } = await import("./extension.js");
-      return markdownExtensions();
-    },
-  },
-  {
     type: "patchwork:datatype",
     id: "markdown",
     name: "Markdown",
@@ -30,6 +20,26 @@ export const plugins = [
     async load() {
       const { MarkdownDatatype } = await import("./datatype.js");
       return MarkdownDatatype;
+    },
+  },
+  {
+    type: "codemirror:extension",
+    id: "codemirror-markdown",
+    name: "Markdown",
+    supportedDatatypes: ["essay", "markdown"],
+    async load(): Promise<Extension> {
+      const { markdownExtensions } = await import("./extensions/markdown.js");
+      return markdownExtensions();
+    },
+  },
+  {
+    type: "codemirror:extension",
+    id: "codemirror-markdown-links",
+    name: "Markdown Clickable Links",
+    supportedDatatypes: ["essay", "markdown"],
+    async load(): Promise<Extension> {
+      const { markdownLinks } = await import("./extensions/links.js");
+      return markdownLinks();
     },
   },
 ];
