--- tools/codemirror/codemirror-md-links/esbuild/build.ts
@@ -1,3 +1,0 @@
-import options from "./options.ts";
-import * as esbuild from "esbuild";
-esbuild.build(options);
