--- tools/codemirror/codemirror-md-links/esbuild/context.ts
@@ -1,3 +1,0 @@
-import options from "./options.ts";
-import * as esbuild from "esbuild";
-export default await esbuild.context(options);
