--- tools/codemirror/codemirror-md-links/esbuild/watch.ts
@@ -1,2 +1,0 @@
-import context from "./context.ts";
-await context.watch();
