--- tools/codemirror/codemirror-md-links/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@grjte/codemirror-md-links",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "main": "./dist/main.js",
-  "exports": {
-    ".": {
-      "import": "./dist/main.js",
-      "source": "./src/main.tsx"
-    },
-    "./extension": {
-      "import": "./dist/extension.js",
-      "source": "./src/extension.ts"
-    }
-  },
-  "scripts": {
-    "build": "node esbuild/build.ts",
-    "dev": "node esbuild/watch.ts",
-    "pushwatch": "pnpm dev pushwork"
-  },
-  "dependencies": {
-    "@codemirror/language": "^6.11.3",
-    "@codemirror/state": "^6.5.2",
-    "@codemirror/view": "^6.38.4",
-    "@inkandswitch/patchwork-bootloader": "workspace:^"
-  },
-  "devDependencies": {
-    "@babel/core": "^7.28.4",
-    "@babel/parser": "^7.28.4",
-    "@eslint/js": "^9.36.0",
-    "@node-types/node": "npm:@types/node@^24.9.1",
-    "@types/babel__core": "^7.20.5",
-    "esbuild": "^0.23.1",
-    "eslint": "^9.36.0",
-    "globals": "^16.4.0",
-    "typescript": "~5.9.3",
-    "typescript-eslint": "^8.45.0"
-  }
-}
