+++ tools/editors/tenfold/src/file-viewer.tsx
@@ -1,7 +1,7 @@
 import type { DocHandle } from "@automerge/automerge-repo";
 import type { PatchworkViewElement } from "@inkandswitch/patchwork-elements";
 import { makeDocumentProjection } from "@automerge/automerge-repo-solid-primitives";
-import { markdownExtensions } from "@grjte/codemirror-markdown/extension";
+import { markdownExtensions } from "@grjte/codemirror-markdown/markdown";
 import { CodeMirror } from "@grjte/codemirror-base/component";
 import { githubLight, githubDark } from "@uiw/codemirror-theme-github";
